Test: Finish test for msg_write_obj_ack

+++ src/message/msg_write_obj.c
@@ -76,7 +76,7 @@ extern int recv_msg_write_obj_ack(int fd, msg_write_obj_ack* msg, int64_t log_id
     int ret = 0;
     CCEPH_RECV_FIELD(client_id, int32, &msg->client_id);
     CCEPH_RECV_FIELD(req_id, int32, &msg->req_id);
-    CCEPH_RECV_FIELD(result, int64, &msg->result);
+    CCEPH_RECV_FIELD(result, int8, &msg->result);
     return 0;
 }
 extern int send_msg_write_obj_ack(int fd, msg_write_obj_ack* msg, int64_t log_id) {
@@ -84,6 +84,6 @@ extern int send_msg_write_obj_ack(int fd, msg_write_obj_ack* msg, int64_t log_id
     int ret = 0;
     CCEPH_SEND_FIELD(client_id, int32, msg->client_id);
     CCEPH_SEND_FIELD(req_id, int32, msg->req_id);
-    CCEPH_RECV_FIELD(result, int64, &msg->result);
+    CCEPH_RECV_FIELD(result, int8, &msg->result);
     return 0;
 }
